requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
nltk
Pillow
scikit-learn
tqdm
pickle-mixin
